refactor(Vsix): match namespace to "Contracts" folder.
