+++ src/LibraryManager.Vsix/Commands/CleanCommand.cs
@@ -6,6 +6,7 @@ using System.ComponentModel.Design;
 using System.Threading;
 using EnvDTE;
 using Microsoft.VisualStudio.Shell;
+using Microsoft.Web.LibraryManager.Vsix.Contracts;
 using Microsoft.Web.LibraryManager.Vsix.Shared;
 using Task = System.Threading.Tasks.Task;
