+++ src/LibraryManager.Vsix/Contracts/Logger.cs
@@ -10,7 +10,7 @@ using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.Logging;
 using Microsoft.Web.LibraryManager.Vsix.Shared;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Contracts
 {
     internal static class Logger
     {
